Adding cublasSgemmGroupedBatched and adding support for float32

+++ grouped_gemm/ops_test.py
@@ -37,7 +37,7 @@ _TEST_PROBLEMS = add_transpose_flags((
 
 def randn(bs, x, y):
     out = (torch.rand(bs, x, y) - 0.5 * 2) / (y * x)
-    return out.cuda().to(torch.half)
+    return out.cuda().to(torch.float32)
 
 
 def gmm(a, b, batch_sizes, trans_b=False):
